add event list and create

+++ app/src/main/java/com/jiayantech/jyandroid/model/DiaryHeader.java
@@ -17,11 +17,11 @@ public class DiaryHeader extends BaseModel {
     public String categoryIds;
     public String projectId;
     public String projectName;
-    public long operationTime;
+    public double operationTime;
     public long price;
     public float satisfyLevel;
     public String tags;
-    public long lastModifyTime;
+    public double lastModifyTime;
     public String previousPhoto;
     public String currentPhoto;
